refactor: 음식점 좋아요 역정규화를 통한 쿼리 개선 (#517)

+++ backend/src/main/java/com/celuveat/restaurant/query/dto/RestaurantWithDistance.java
@@ -11,6 +11,6 @@ public record RestaurantWithDistance(
         String naverMapUrl,
         Integer viewCount,
         Double distance,
-        Long likeCount
+        int likeCount
 ) {
 }
